fix build on 32-bit.

+++ muduo/base/FileUtil.cc
@@ -60,7 +60,7 @@ int FileUtil::SmallFile::readToString(int maxSize,
         if (S_ISREG(statbuf.st_mode))
         {
           *fileSize = statbuf.st_size;
-          content->reserve(std::min(implicit_cast<int64_t>(maxSize), *fileSize));
+          content->reserve(static_cast<int>(std::min(implicit_cast<int64_t>(maxSize), *fileSize)));
         }
         else if (S_ISDIR(statbuf.st_mode))
         {
